Moving HW
This commit adds the moving hw. The project will open a window in open
gl and will render the characters and fruits along with a background.
The character will move left and right using A and D Key. The fruits
will fall and when they disappear off the screen they will start back at
the top.