Collision HW
This commit adds in the collision HW. In this project I added in the
collision detection to the engine.  I added the game Object class to
hold and control the game objects. Each game object has a transform,
rigidBody struct and a collider state. In this project the game will run
in engine and detect collisions between the fruits and player.

Side Note: Sphere and AABB collisons are a little funky and have some
bugs. Character can also move using arrow keys.